WIP: refactored 'onClickLink', 'onEditLink', 'onRemoveLink', 'renderLinkButton', 'renderLinkDialogWindow'
